Develop -> Main

+++ server/src/main/java/com/example/TicketGuru/web/VenueRestContoller.java
@@ -1,58 +1,88 @@
 package com.example.TicketGuru.web;
 
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
-import com.example.TicketGuru.domain.Event;
+import com.example.TicketGuru.domain.PostalCode;
+import com.example.TicketGuru.domain.PostalCodeRepository;
 import com.example.TicketGuru.domain.Venue;
 import com.example.TicketGuru.domain.VenueRepository;
+
+import jakarta.validation.Valid;
 
 @RestController
 public class VenueRestContoller {
 
 	@Autowired
 	VenueRepository venueRepository;
-	
-	// palauttaa listan tapahtumapaikoista
+	@Autowired
+	PostalCodeRepository pcrepository;
+
+	// Palauttaa listan tapahtumapaikoista
 	@GetMapping("/venues")
-	public Iterable<Venue> getVenues(){
+	public Iterable<Venue> getVenues() {
+		List<Venue> venues = (List<Venue>) venueRepository.findAll();
 		return venueRepository.findAll();
 	}
-	
+
 	// Palauttaa tapahtumapaikan id:llä
 	@GetMapping("/venues/{venueId}")
 	public Optional<Venue> getVenue(@PathVariable("venueId") Long venueId) {
+		Optional<Venue> venue = venueRepository.findById(venueId);
+		if (venue.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Annetulla id:llä ei löytynyt tapahtumapaikkaa");
+		}
 		return venueRepository.findById(venueId);
 	}
-	
-	// lisää uuden tapahtumapaikan
+
+	// Lisää uuden tapahtumapaikan
 	@PostMapping("/venues")
-	Venue newVenue(@RequestBody Venue newVenue) {
+	@ResponseStatus(HttpStatus.CREATED)
+	public Venue newVenue(@Valid @RequestBody Venue newVenue) {
+		Optional<PostalCode> postcode = pcrepository.findByPostalCode(newVenue.getPostalCode().getPostalCode());
+		if (postcode.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Annettua postinumeroa ei löydy");
+		}
 		return venueRepository.save(newVenue);
 	}
-	
-	// muokkaa tapahtumapaikkaa, jolla valittu venueId
-	// jos tapahtumapaikkaa ei löydy ID:llä, metodi luo uuden tapahtuman seuraavalla vapaalla id:llä
+
+	// Muokkaa tapahtumapaikkaa, jolla valittu venueId
 	@PutMapping("/venues/{venueId}")
-	Venue replaceVenue(@RequestBody Venue editedVenue, @PathVariable("venueId") Long venueId) {
-		editedVenue.setVenueId(venueId);
-		return venueRepository.save(editedVenue);
+	public Venue editVenue(@Valid @RequestBody Venue editedVenue, @PathVariable("venueId") Long venueId) {
+		Optional<Venue> venue = venueRepository.findById(venueId);
+		if (venue.isPresent()) {
+			try {
+				editedVenue.setVenueId(venueId);
+				return venueRepository.save(editedVenue);
+			} catch (Exception e) {
+				throw new ResponseStatusException(HttpStatus.BAD_REQUEST, e.getMessage());
+			}
+		} else {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Annetulla id:llä ei löytynyt tapahtumapaikkaa");
+		}
 	}
-	//Poistaa tapahtumapaikan id:n perusteella
+
+	// Poistaa tapahtumapaikan id:n perusteella
 	@DeleteMapping("/venues/{venueId}")
 	public Iterable<Venue> deleteVenue(@PathVariable("venueId") Long venueId) {
+		Optional<Venue> venue = venueRepository.findById(venueId);
+		if (venue.isEmpty()) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Annetulla id:llä ei löytynyt tapahtumapaikkaa");
+		}
 		venueRepository.deleteById(venueId);
 		return venueRepository.findAll();
-		}
-		
+	}
+
 }
